Remove redundant files in groups section
