+++ features/admin.groups.v1/components/wizard/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2020-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,7 +16,4 @@
  * under the License.
  */
 
-export * from "./group-basics";
-export * from "./group-assign-users";
 export * from "./group-summary";
-export * from "./create-group-wizard";
